requirements.txt
PyYAML
requests
jwcrypto
pypng
